refactor: split minimap, diplomacy chrome, region map, and fallback runner parts (#3878)
Decompose region minimap zoom/toggle chrome, grant/subsidy dialog chrome,
CtRegionMap bus/tap handlers, and GameScreen fallback next-turn resolution
into library parts without changing public APIs or player-facing behavior.

Refs #3878
