+++ tool/check_app_event_bus_decoupling.dart
@@ -86,7 +86,7 @@ const Set<String> _allowedFeatureLocalDialogFiles = <String>{
   // modularization. Same local-by-design rationale as the parent file
   // (Refs #3878 Phase 3).
   'app/lib/features/game/screens/game/game_screen_fallback_next_turn.dart',
-  'app/lib/features/game/screens/game/game_screen_fallback_next_turn_resolution.dart',
+  'app/lib/features/game/screens/game/game_screen_fallback_next_turn_runner.dart',
   // Android back / exit-to-main-menu confirm dialog extracted from
   // game_screen.dart per `SPEC/ui/in-game-shell-narrow.md` "Android back
   // confirm". Local by design — `SPEC/program/app-ui-wiring.md` line 84.
